requirements.txt
# Core dependencies
google-adk[eval]  # Includes eval module for evaluation features
google-genai
pypdf
requests
python-dotenv  # For .env file support

# Google Slides API
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Google Cloud Vision API (for layout review)
google-cloud-vision>=3.0.0

# PDF to image conversion
pdf2image>=1.16.0
Pillow>=10.0.0

# Chart generation (Plotly - best UI, free, can export static PNG)
plotly>=5.18.0
kaleido>=0.2.1  # Required for static image export (PNG/JPEG)

# HTTP server for Cloud Run
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0
nest-asyncio>=1.6.0

# Development dependencies (optional)
ipython

# Open Peeps illustration search
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
fastapi>=0.104.0
uvicorn>=0.24.0

